:recycle: refactor: remove save facade method and move its logic into save vote usecase
