+++ kok-core/src/main/java/com/kok/kokcore/vote/usecase/SaveVoteUseCase.java
@@ -1,9 +1,8 @@
 package com.kok.kokcore.vote.usecase;
 
-import com.kok.kokcore.vote.domain.Vote;
 import java.util.List;
 
 public interface SaveVoteUseCase {
 
-    void saveVotes(List<Vote> votes);
+    void saveVotes(String roomId, String memberId, List<Long> agreedStationIds);
 }
